Цена при результате должна быть float

+++ src/main/java/com/foxelyss/transportbooking/repos/TransportingResult.java
@@ -4,6 +4,6 @@ import java.sql.Timestamp;
 
 public record TransportingResult(int id, String name, Timestamp start, Timestamp end,
                                  String startPoint, String endPoint,
-                                 int arr, int dep, double price, String mean,
+                                 int arr, int dep, float price, String mean,
                                  String company, int places, int freePlaceQuantity) {
 }
